smart-expense-tracker/ai-service/requirements.txt
fastapi
uvicorn[standard]
easyocr
pillow
spacy
python-dateutil
scikit-learn
joblib
python-multipart
numpy
